Move `CommitForm` file outside of `OutlineTree`
