+++ apps/lite/ui/src/routes/project/$id/workspace/OutlineTree/UncommittedChangesRow.tsx
@@ -22,7 +22,7 @@ import { RowBubble, RowBubbleGroup, RowLabel, RowLabelContainer, RowToolbar } fr
 import { ItemRow } from "./ItemRow.tsx";
 import { useQueries } from "@tanstack/react-query";
 import { treeChangeDiffsQueryOptions } from "#ui/api/queries.ts";
-import { commitMessageInputId } from "./CommitForm.tsx";
+import { commitMessageInputId } from "../CommitForm.tsx";
 
 type LineStats = {
 	linesAdded: number;
